project/pyproject.toml
[tool.poetry]
name = "restapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask-Bootstrap = "^3.3.7"
arrow = "^1.1.1"
boto3 = "^1.18.48"
Flask-RESTful = "^0.3.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
py-healthcheck = "^1.10.1"
flake8 = "^3.9.2"
black = "^21.9b0"
mypy = "^0.910"
bandit = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
